+++ src/js/templates/forms/updatePostForm.js
@@ -1,6 +1,4 @@
-import { getPostFromAPI } from "../../api/requests/index.js";
 import { updatePostContent } from "../../ui/events/index.js";
-
 
 export function updateForm(id) {
   const form = document.createElement("form");
@@ -23,31 +21,8 @@ export function updateForm(id) {
   updateBtn.innerText = "update post";
   updateBtn.type = "submit";
 
-  form.append(updateBtn );
+  form.append(updateBtn);
   form.addEventListener("submit", updatePostContent);
 
   return form;
-}
-
-export async function populateUpdateForm(updateForm) {
-  const id = updateForm.name
-
-  if (updateForm) {
-    const button = updateForm.querySelector("button");
-    button.disabled = true;
-
-    const postPopulation = await getPostFromAPI(id);
-
-    if (postPopulation.title) {
-      updateForm.title.value = postPopulation.title;
-    }
-    if (postPopulation.body) {
-      updateForm.body.value = postPopulation.body;
-    }
-    if (!postPopulation.body || !postPopulation.title) {
-      console.log("the post is lacking some content");
-    }
-
-    button.disabled = false;
-  }
 }
