+++ src/js/templates/modals/renderModalBody.js
@@ -1,7 +1,8 @@
 import { removePostFromAPI } from "../../api/requests/index.js";
 import { load } from "../../storage/index.js";
-import { removeUrlParameter } from "../../ui/listeners/index.js";
-import { populateUpdateForm, updateForm } from "../forms/index.js";
+import { populateUpdateForm } from "../../ui/events/index.js";
+import { removeUrlParameter } from "../../ui/urlParams/index.js";
+import { updateForm } from "../forms/index.js";
 
 const author = load("profile");
 
@@ -21,7 +22,7 @@ export function renderModalBody(postData) {
     const form = updateForm(postData.id);
     modalBody.append(form);
     populateUpdateForm(form);
-    
+
     const removePost = makeDeleteBtn();
     modalBody.append(removePost);
 
@@ -29,7 +30,6 @@ export function renderModalBody(postData) {
       await removePostFromAPI(postData.id);
       removeUrlParameter("post-id");
     });
-
   } else if (author.name !== postData.author.name) {
     if (postData.media) {
       const img = document.createElement("img");
