+++ src/js/ui/events/index.js
@@ -1,6 +1,7 @@
 export * from "./auth.js";
 export * from "./newPost.js";
-export * from "./updatePostContent.js";
 export * from "./search.js";
 export * from "./filter.js";
+export * from "./updatePostContent.js";
+export * from "./populateUpdateForm.js";
