+++ src/js/ui/listeners/index.js
@@ -1,7 +1,8 @@
 export * from "./onAuth.js";
+export * from "./openModal.js"
+export * from "./removeModal.js"
 export * from "./publishPost.js";
 export * from "./onSearch.js"
-export * from "./openModal.js"
 export * from "./onFilter.js"
 export * from "./logOut.js"
