+++ src/js/ui/urlParams/index.js
@@ -1,0 +1,2 @@
+export * from "./insertUrlParam.js"
+export * from "./removeUrlParam.js"
